requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
Django==5.1.3
et_xmlfile==2.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
openpyxl==3.1.5
pillow==11.0.0
reportlab==4.2.5
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.2
urllib3==2.2.3
